Implemented review comments.
